refactor: mundando algumas classes de lugares
