+++ ms-avaliador-credito/src/main/java/br/com/msavaliadorcredito/domian/model/cliente/DadosCliente.java
@@ -1,4 +1,4 @@
-package br.com.msavaliadorcredito.domian.model;
+package br.com.msavaliadorcredito.domian.model.cliente;
 
 import lombok.Data;
