+++ ms-avaliador-credito/src/main/java/br/com/msavaliadorcredito/domian/model/dados/RetornoAvaliacaoCliente.java
@@ -1,5 +1,6 @@
-package br.com.msavaliadorcredito.domian.model;
+package br.com.msavaliadorcredito.domian.model.dados;
 
+import br.com.msavaliadorcredito.domian.model.cartao.CartaoAprovado;
 import lombok.AllArgsConstructor;
 import lombok.Data;
